update the example for vue
