+++ npm/vue/examples/cli-ts/cypress/plugins/index.ts
@@ -1,7 +1,8 @@
 const preprocessor = require('@cypress/vue/dist/plugins/webpack')
+const webpackConfig = require('@vue/cli-service/webpack.config')
 
 module.exports = (on, config) => {
-  preprocessor(on, config)
+  preprocessor(on, config, webpackConfig)
 
   // IMPORTANT return the config object
   return config
