--- npm/vue/src/preprocessor/index.js
@@ -1,3 +1,0 @@
-const onFilePreprocessor = require('./webpack').onFilePreprocessor
-
-module.exports = onFilePreprocessor
