+++ npm/webpack-dev-server/src/makeWebpackConfig.ts
@@ -3,7 +3,7 @@ import * as path from 'path'
 import { merge } from 'webpack-merge'
 import CypressCTOptionsPlugin from './plugin'
 
-const debug = debugFn('cypress:ct:webpack')
+const debug = debugFn('@cypress/webpack-dev-server:makeWebpackConfig')
 
 export async function makeWebpackConfig (userWebpackConfig = {}, { projectRoot, files, support }) {
   debug(`User passed in webpack config with values`, userWebpackConfig)
